test_requirements.txt
-e .
colour-runner==0.0.4
coverage==3.7.1
dj-database-url==0.3
django==1.7.0
django-sirtrevor==0.2.3
factory_boy==2.4.1
incuna-test-utils==4.0.0
psycopg2==2.5.4
